created a separate debounce hook
