+++ frontend/src/hooks/useDebounce.ts
@@ -1,0 +1,15 @@
+import { useState, useEffect } from 'react';
+function useDebounce<T>(value: T, delay: number): T {
+  const [debounceValue, setDebounceValue] = useState(value);
+
+  useEffect(() => {
+    const timeout = setTimeout(() => {
+      setDebounceValue(value);
+    }, delay);
+
+    return () => clearTimeout(timeout);
+  }, [value, delay]);
+
+  return debounceValue;
+}
+export default useDebounce;
